Use more descriptive libvirt secret names corresponding to reality
This patch changes the name for libvirt secrets, as the user might not
necessarily use the 'nova' and 'cinder' usernames, but completely
different names. The names 'client.nova' and 'client.cinder' might
suggest to the user that they are templatable, but they are not.
In Kolla, there can only be one secret for the nova service and one
for cinder, regardless of how many clusters are being used.
Therefore, I use a name that does not resemble the Ceph keyring to
prevent this from happening. A description has also been added.

Change-Id: Id80e45357ea8054f2bacb8a7c687b335c74d5feb