Set lock_path for openstack services
The Oslo.messaging project implemented a nice feature
called Queue Manager. This means that instead of using
random queue names in RabbitMQ, it uses queue names composed
of hostname, process-name and some integer. For proper
functioning, the code using lockutils from oslo_concurrency
and it's failing if lock_path is not set.

This patch simply set lock_path of oslo_concurrency for
services which can potentionally use Queue Manager.

This is enabled when ``om_enable_queue_manager`` is set to ``True``.
Queue Manager will be supported in a follow-up patch.

Change-Id: Ided3b2bce03ea11fb34820fc40b4a3d694b8b44c