[yoga-only] Cap setuptools <20.26.4
py39 jobs (on ubuntu-focal) started to fail due to recent setuptools
release (20.26.4) on Yoga, because we have 'packaging==21.3' in this
branch that is not compatible with the new setuptools [1].

setuptools is bundled in virtualenv, so it has to be capped via the
virtualenv package. tox also needed to be capped (<4) as gate uses
tox 3.28.0, but with capping virtualenv we pull in latest tox as well,
which would cause other errors.

[1] https://github.com/pypa/setuptools/issues/4483

Change-Id: I49deaf07aa3f50a8bb9e3191c864277990ac4dda