WIP: Add ovn-bgp-agent / FRR / neutron-bgpvpn / Horizon BGPVPN dashboard support
ovn-bgp-agent [1] is a new component that exposes VMs/containers through BGP on OVN environments.

This change extends kolla-ansible to deploy ovn-bgp-agent and FRR, as this is a requirement for
ovn-bgp-agent.
The neutron-server is extended to be able to enable the networking-bgpvpn plugin.
The openvswitch deployment is modified to allow for br-ex without connected physical ports as
the traffic is routed instead.
Furthermore, it also extends the horizon integration to enable the BGPVPN dashboard plugin,
that is also required for certain scenarios with ovn-bgp-agent.

This change is part of a set of changes in several projects that are nessecary to fully
implement ovn-bgp-agent.

Depends-On: https://review.opendev.org/c/openstack/kolla/+/891617

[1] https://opendev.org/openstack/ovn-bgp-agent

Change-Id: I4fde95dccd036e371239e21596c05ecfd5342270