Add key type for letsencrypt certificate
Allows the insertion of a key type for letsencrypt, due to the
fact that there are other 3rd party solutions (i.e. Vault) that
require specific RSA key types.

This change allows the ability to use vault instead of the public
DNS so we can effectively provide private DNS certificates
to an environment.

Depends-On: https://review.opendev.org/c/openstack/kolla/+/934681
Closes-Bug: 2087860
Change-Id: I9ba49b1150e656dcb16ee3cf3b79b56e7ef2d3a5