docs: minor fixes to external Ceph guide
When a path is broken into two separate lines, the generated HTML
includes an incorrect space character in the middle of the path.

We need to ignore doc8 linting on this file as this change breaks the
max line length check. I tried using ``ignore-path-errors`` to exclude
only D001 for this file, but doc8 does not like a relative path for this
option [1].

[1] https://github.com/PyCQA/doc8/issues/162

Change-Id: I1cf0b7a6b7506a68876bce2de5080718280a47be