Fix unrestricted copying of backend TLS certs
With the change
https://review.opendev.org/c/openstack/kolla-ansible/+/915901
K-A copies backend TLS certificates to all containers of projects
when backend TLS for the project is enabled because of less strict
conditional check.

However, this causes K-A to attempt copying non-existent certificates
from hosts that is not part of backend TLS to containers that are
also not part of backend TLS.

This change makes service-copy-certs task check if a host is part of
``tls-backend`` group when copying backend TLS certificate and key.

Cloeses-Bug: #2105505
Change-Id: I799dc6e6fdccc483784fc3c2088ada69f24412e9
Signed-off-by: Seunghun Lee <seunghun@stackhpc.com>
(cherry picked from commit e092d33245828b16c07fa4364a91e903dcbe1de0)