[CI] Fix bootstrap testing
This patch adjusts the CI in such a way that we
can finally test the correctness of bootstrap-servers.

 - It correctly installs the virtualenv at the beginning
 - It sets ansible_python_interpreter to this virtualenv
 - It runs bootstrap-servers between upgrades
 - It moves package installation from the CI playbook
   to ansible-collections-kolla

To summarize: this patch aligns behavior and tests the code
within the boundaries it should be tested in - that is, inside
a virtualenv. This naturally also tests the correctness of
ansible-collection-kolla, which installs packages and other
dependencies (now not mixed with OS py interpreter).
This is important, especially due to a bug in older versions
of Docker and requests, which required running bootstrap-servers.
Proof of this is that the rocky9-octavia job is finally working,
with no more issues with package cryptography (which was read from OS
python environment by ansible).

TL;DR: From now on, it runs correctly inside a virtualenv,
and the OS interpreter no longer interferes.

Depends-On: https://review.opendev.org/c/openstack/ansible-collection-kolla/+/946983

Change-Id: Ifea75c2e14ffececbe271bee1549e92bded90d48