[CI] Enable virtualenv testing for Debian/Ubuntu
This patch enables testing with a runtime virtualenv on Debian and Ubuntu
in CI jobs and without for RHEL derivatives. For upgrade scenarios, where
the previous stable branch is checked out and does not yet support
automatic venv creation, a temporary workaround is added to create the
venv manually in the Zuul playbook.

Summary:
- New deployments on Debian/Ubuntu now test virtualenv creation logic.
- Upgrade jobs set ansible_python_interpreter only if testing a previous
  release, and a manual venv is created in the playbook as a workaround.
- This workaround is only needed until the following patches are
  merged into all relevant stable branches.

The workaround can be removed once these are merged:

[1] https://review.opendev.org/c/openstack/kolla-ansible/+/949813
[2] https://review.opendev.org/c/openstack/ansible-collection-kolla/+/950265

Change-Id: I5d4c5d960441025da1fc1b6f197b03476b1be1a9