[letsencrypt] Avoid rendering empty --key-type argument
The `--key-type` argument was always rendered in the
command, even if the `letsencrypt_key_type` variable was
empty. This caused the command to fail due to an invalid
empty argument.

The template now checks if `letsencrypt_key_type` has a
non-zero length before including `--key-type`, ensuring correct
rendering of the script.

Signed-off-by: Michal Arbet <michal.arbet@ultimum.io>

Change-Id: I66a76fc8d27bcc97cd3133eed137fd46ea599511