Move OpenSearch role right after loadbalancer
After fluentd CI log checks have been implemented, in the only scenario
that deploys log management stack (prometheus-opensearch) we're seeing
issues that fluentd is not reconnecting to OpenSearch deployed late in
the workflow.
It's not reconnecting fast enough, and if the connection is not there,
fluentd is not parsing/opening log files - which we're checking for.

Change-Id: I1f4839676b4f794bb93f1372aa16f8377ae9b269
Signed-off-by: Michal Nasiadka <mnasiadka@gmail.com>